devide `addOrder.ts`
